Add dpp bash shellscript
dpp is a 24-line bash shellscript that takes one input parameter, a cpp conditional.  It processes the document, selecting any paragraph that matches the conditional and rejecting any #ifdef'ed sections with non-matching conditionals.  Also makes sure to remove all conditionals so that no stray control text is left in the finished document.  Output the desired document to stdout.